Add support for SHN_COMMON.
The SHN_COMMON pseudo-section is used in the symbol tables to reference
undefined global variables. If all instances of the symbol are in the
pseudo-section SHN_COMMON, space is reserved in .bss. If the variable
is actualy defined, and ends up as GLOBAL in one object, no space is
reserved in .bss, and the defined instance is used.